Move implementation of AsciiDoc line to Line
