+++ redpen-core/src/main/java/cc/redpen/parser/asciidoc/AsciiDocLine.java
@@ -18,16 +18,13 @@
 
 package cc.redpen.parser.asciidoc;
 
-import java.util.ArrayList;
-import java.util.List;
+import cc.redpen.parser.common.Line;
 
 
 /**
  * An 'erasing' string utility class that stores the original offset for each preserved character
  */
-public class AsciiDocLine {
-    // value returned for comparison if a character is escaped
-    static final char ESCAPED_CHARACTER_VALUE = 'ø';
+public class AsciiDocLine extends Line {
     static final String INLINE_MARKUP_DELIMITERS = " _*`#^~.,";
 
     /**
@@ -42,18 +39,9 @@ public class AsciiDocLine {
         CloseMarkerContainsDelimiters
     }
 
-    // a list of offsets for each character
-    List<Integer> offsets = new ArrayList<>();
-    // the text for the line
-    List<Character> text = new ArrayList<>();
-    // marks erased characters as invalid
-    List<Boolean> valid = new ArrayList<>();
-    // remembers which characters were escaped in the original string
-    List<Boolean> escaped = new ArrayList<>();
 
     private int lineNo = 0;
     private boolean allSameCharacter = false;
-    private boolean erased = false;
     private boolean inBlock = false;
 
     private int sectionLevel = 0;
@@ -174,51 +162,6 @@ public class AsciiDocLine {
     }
 
     /**
-     * Erase length characters in the line, starting at pos
-     *
-     * @param pos start position
-     * @param length length to erase
-     */
-    public void erase(int pos, int length) {
-        if ((pos >= 0) && (pos < valid.size())) {
-            for (int i = pos; (i < valid.size()) && (i < pos + length); i++) {
-                valid.set(i, false);
-            }
-        }
-    }
-
-    /**
-     * Erase the whole line
-     */
-    public void erase() {
-        for (int i = 0; i < valid.size(); i++) {
-            valid.set(i, false);
-        }
-        erased = true;
-    }
-
-    /**
-     * Erase all occurrences of the given string
-     *
-     * @param segment segment to be erased
-     */
-    public void erase(String segment) {
-        for (int i = 0; i < text.size(); i++) {
-            boolean found = true;
-            for (int j = 0; j < segment.length(); j++) {
-                if (charAt(j + i) != segment.charAt(j)) {
-                    found = false;
-                    break;
-                }
-            }
-            if (found) {
-                erase(i, segment.length());
-                i += segment.length();
-            }
-        }
-    }
-
-    /**
      * Erase the open and close markers, and optionally all the text inside them
      * Returns the position of the first enclosure or -1 if no enclosure was found
      *
@@ -331,127 +274,12 @@ public class AsciiDocLine {
     }
 
     /**
-     * Return the length of the line
-     *
-     * @return length of the line
-     */
-    public int length() {
-        return text.size();
-    }
-
-    /**
-     * Return the character at the given position. Erase characters will return 0 rather
-     * than the actual character
-     *
-     * @param i index
-     * @return extracted character
-     */
-    public char charAt(int i) {
-        return charAt(i, false);
-    }
-
-    /**
-     * Return the character at the given position, optionally including erased characters
-     *
-     * @param i index
-     * @param includeInvalid true if include invalid
-     * @return extracted character
-     */
-    public char charAt(int i, boolean includeInvalid) {
-        if ((i >= 0) && (i < text.size())) {
-            if (escaped.get(i)) {
-                return ESCAPED_CHARACTER_VALUE;
-            }
-            if (includeInvalid || valid.get(i)) {
-                return text.get(i);
-            }
-        }
-        return 0;
-    }
-
-    /**
-     * Return the offset for the character at the given position
-     *
-     * @param i index
-     * @return offset
-     */
-    public int getOffset(int i) {
-        if (i >= 0) {
-            if (i < offsets.size()) {
-                return offsets.get(i);
-            }
-            else {
-                return offsets.size();
-            }
-        }
-        return 0;
-    }
-
-    /**
      * Return the original line number for this line
      *
      * @return line number
      */
     public int getLineNo() {
         return lineNo;
-    }
-
-    /**
-     * Return the raw character at the specified position, ignoring its validity.
-     *
-     * @param i index
-     * @return raw character at the specified position
-     */
-    public char rawCharAt(int i) {
-        if ((i >= 0) && (i < text.size())) {
-            return text.get(i);
-        }
-        return ' ';
-    }
-
-    /**
-     * Is the character at the given position valid?
-     *
-     * @param i index
-     * @return true if the character at the given position valid
-     */
-    public boolean isValid(int i) {
-        if ((i >= 0) && (i < text.size())) {
-            return valid.get(i);
-        }
-        return false;
-    }
-
-    /**
-     * Is the character at the given position empty (ie: whitespace or invalid)
-     *
-     * @return true if it's empty
-     */
-    public boolean isEmpty() {
-        for (int i = 0; i < text.size(); i++) {
-            if (!Character.isWhitespace(text.get(i)))
-                if (valid.get(i)) {
-                    return false;
-                }
-        }
-
-        return true;
-    }
-
-    /**
-     * Does the line start with the given string?
-     *
-     * @param s string to test
-     * @return true if the string starts with the specified string
-     */
-    public boolean startsWith(String s) {
-        for (int i = 0; i < s.length(); i++) {
-            if (charAt(i) != s.charAt(i)) {
-                return false;
-            }
-        }
-
-        return true;
     }
 
     /**
